requirements.txt
fastapi
uvicorn
pydantic
starlette
click
anyio
sniffio
websockets
python-multipart
twilio
python-dotenv
pydub
python-dateutil
six
httpx
pygments
rich
attrs
openai>=1.86.0,<2
PyJWT
tzdata
pinecone
langchain-openai>=0.3.0
langchain-text-splitters
